+++ NEWS.md
@@ -1,4 +1,11 @@
-* Added support for usethis' UI
+* Added `write_bibliography()` to write a bibliography on disk, with
+  options to split the file accordingly to the number of references into
+  each one;
+
+* Added `search_on_pubmed()` to retrieve a bibliography from a search 
+  on PubMed, with option of select the amount of records to retrieve;
+
+* Added support for usethis' UI.
 
 # bibsearch (development version)